rustcodepro
a encoderb approach based on the spare kmer to classify the vairant associated sequences. Why kmers because i can easily loop them into the why sparesekmer, as i can easily make a graph structure based on the loop and use the weight of the loop as 2 to add to the weighted kmer for the representation.